+++ CHANGELOG.rst
@@ -3,6 +3,9 @@ Membranecurvature CHANGELOG
 
 ---
 
+* Added ``curvature_average`` parameter to choose between time-averaged per-frame
+  curvatures (``'per_frame'``, default) and curvature of the time-averaged surface
+  (``'surface'``)
 * Added ``surface_method='binning_nearest'`` as a new surface method (PR #246)
 * Feature: Added periodic edge padding for the binning surface method (PR #238)
 * Refactored binning surface method to use ``np.add.at`` for improved performance (PR #240)